modularize benchmarking scripts
