+++ tests/perf/output_benchmarks20230221-162052.csv
@@ -1,0 +1,4 @@
+Op,ObjectSize,AppBufferSize,LibBufferSize,Crc32cEnabled,MD5Enabled,ApiName,ElapsedTimeUs,CpuTimeUs,Status
+TM_WRITE,10858,-1,104857600,False,False,JSON,453678,-1,['OK']
+TM_WRITE,6091,-1,104857600,False,False,JSON,430287,-1,['OK']
+TM_WRITE,8656,-1,104857600,False,False,JSON,429899,-1,['OK']
